NetworkServices: Implement support for communicating over the R3 internal network service in driverless mode, bugref:10297 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@97075 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/NetworkServices/NetLib/IntNetIf.h
@@ -52,7 +52,7 @@ typedef INTNETIFCTX *PINTNETIFCTX;
 
 
 DECLHIDDEN(int) IntNetR3IfCtxCreate(PINTNETIFCTX phIfCtx, const char *pszNetwork, INTNETTRUNKTYPE enmTrunkType,
-                                    const char *pszTrunk, size_t cbSend, size_t cbRecv, uint32_t fFlags);
+                                    const char *pszTrunk, uint32_t cbSend, uint32_t cbRecv, uint32_t fFlags);
 DECLHIDDEN(int) IntNetR3IfCtxDestroy(INTNETIFCTX hIfCtx);
 DECLHIDDEN(int) IntNetR3IfCtxQueryBufferPtr(INTNETIFCTX hIfCtx, PINTNETBUF *ppIfBuf);
 DECLHIDDEN(int) IntNetR3IfCtxSetActive(INTNETIFCTX hIfCtx, bool fActive);
